requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==23.0.0
httpx==0.19.0
pymongo==4.6.0
pydantic==2.5.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0.post1
python-dotenv==1.0.0
motor==3.3.2